requirements.txt
absl-py==1.0.0
altgraph==0.17.2
apturl==0.5.2
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.1.5
attrs==19.3.0
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
cachetools==5.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.28
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
demjson==2.2.4
distro==1.4.0
distro-info===0.23ubuntu1
dlib==19.22.1
entrypoints==0.3
et-xmlfile==1.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gTTS==2.2.4
h11==0.12.0
h5py==3.6.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
imageio==2.13.5
importlib-metadata==4.11.3
jdcal==1.0
jtutils==0.0.8
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
leven==1.0.4
libclang==13.0.0
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Markdown==3.3.6
matplotlib==3.1.2
more-itertools==4.2.0
netifaces==0.10.4
networkx==2.6.3
nose==1.3.7
numexpr==2.7.1
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==0.25.3
pexpect==4.6.0
Pillow==9.0.1
playsound==1.3.0
pluggy==0.13.0
protobuf==3.20.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pytesseract==0.3.9
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.5
python-csv==0.0.13
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
PyWavelets==1.2.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
scikit-image==0.19.1
scipy==1.7.3
SecretStorage==2.3.1
selenium==4.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==1.9.5
systemd-python==234
tables==3.6.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tesseract==0.1.3
tf-estimator-nightly==2.8.0.dev2021122109
thread6==0.2.0
tifffile==2021.11.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.1.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.7
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.1.1
wrapt==1.14.0
wsproto==1.0.0
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==1.0.0
